WAG-432: removed password requirement from sandbox to prod env.
