+++ website/app/role_switcher/views.py
@@ -5,10 +5,6 @@ from django.contrib.auth.models import Group
 from django.contrib.auth import get_user_model
 from django.contrib import messages
 from .forms import RoleSwitchForm, RevertRoleForm
-from django.conf import settings
-from django.contrib.auth.forms import AuthenticationForm
-from django.contrib.auth.views import LoginView
-from django.http import Http404
 
 User = get_user_model()
 
@@ -129,27 +125,3 @@ def switch_role_view(request):
     }
 
     return render(request, "switch_role.html", context)
-
-
-class LocalLoginView(LoginView):
-    template_name = "wagtailadmin/local_login.html"
-    authentication_form = AuthenticationForm
-    redirect_authenticated_user = True
-
-    def dispatch(self, request, *args, **kwargs):
-        if not getattr(settings, "ENABLE_LOCAL_LOGIN", False):
-            raise Http404()
-        configured_tokens = {
-            t
-            for t in [
-                getattr(settings, "LOCAL_LOGIN_TOKEN", ""),
-                getattr(settings, "CYPRESS_LOCAL_LOGIN_TOKEN", ""),
-            ]
-            if t
-        }
-        token = request.GET.get("token", "")
-        if configured_tokens and token not in configured_tokens:
-            raise Http404()
-        resp = super().dispatch(request, *args, **kwargs)
-        resp["X-Robots-Tag"] = "noindex, nofollow, noarchive"
-        return resp
