+++ website/app/settings/base.py
@@ -117,12 +117,7 @@ TEMPLATES = [
     },
 ]
 
-# ------------------------------
-# SSO (Azure) + hidden local-login split
-# ------------------------------
-
-# Env flags for the hidden local (emergency) login page. Keep disabled in prod.
-ENABLE_LOCAL_LOGIN = os.getenv("ENABLE_LOCAL_LOGIN", "false").lower() == "true"
+ENABLE_LOCAL_LOGIN = os.getenv("ENABLE_LOCAL_LOGIN", True)
 LOCAL_LOGIN_TOKEN = os.getenv("LOCAL_LOGIN_TOKEN", "")
 CYPRESS_LOCAL_LOGIN_TOKEN = os.getenv("CYPRESS_LOCAL_LOGIN_TOKEN", "")
